refactor(zoe): move startInstance into separate file
